+++ config/config.go
@@ -255,6 +255,7 @@ func WithTLSSInsecureSkipVerify() Option {
 	}
 }
 
+// WithGrpcOptions appends custom grpc dial options to defaults
 func WithGrpcOptions(option ...grpc.DialOption) Option {
 	return func(c *Config) {
 		c.grpcOptions = append(c.grpcOptions, option...)