+++ internal/conn/conn.go
@@ -4,12 +4,16 @@ import (
 	"context"
 	"fmt"
 	"sync"
+	"sync/atomic"
 	"time"
 
-	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/connectivity"
+	"google.golang.org/grpc/stats"
 
+	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
+
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/backoff"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/endpoint"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/response"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
@@ -46,15 +50,17 @@ func (c *conn) Address() string {
 }
 
 type conn struct {
-	mtx       sync.RWMutex
-	config    Config // ro access
-	cc        *grpc.ClientConn
-	done      chan struct{}
-	endpoint  endpoint.Endpoint // ro access
-	closed    bool
-	state     State
-	lastUsage time.Time
-	onClose   []func(*conn)
+	mtx               sync.RWMutex
+	config            Config // ro access
+	grpcDialOptions   []grpc.DialOption
+	cc                *grpc.ClientConn
+	done              chan struct{}
+	endpoint          endpoint.Endpoint // ro access
+	closed            bool
+	state             State
+	lastUsage         time.Time
+	onClose           []func(*conn)
+	onTransportErrors []func(ctx context.Context, cc Conn, cause error)
 }
 
 func (c *conn) Ping(ctx context.Context) error {
@@ -197,19 +203,40 @@ func (c *conn) take(ctx context.Context) (cc *grpc.ClientConn, err error) {
 		defer cancel()
 	}
 
-	// prepend "ydb" scheme for grpc dns-xresolver to find the proper scheme
-	// "ydb:///", three slashes is ok. It needs for good parse scheme in grpc resolver.
+	// prepend "ydb" scheme for grpc dns-resolver to find the proper scheme
+	// three slashes in "ydb:///" is ok. It needs for good parse scheme in grpc resolver.
 	address := "ydb:///" + c.endpoint.Address()
 
-	cc, err = grpc.DialContext(ctx, address, c.config.GrpcDialOptions()...)
+	cc, err = grpc.DialContext(
+		ctx,
+		address,
+		c.grpcDialOptions...,
+	)
 	if err != nil {
-		return nil, xerrors.WithStackTrace(fmt.Errorf("dial %s failed: %w", address, err))
+		err = xerrors.WithStackTrace(
+			xerrors.Retryable(
+				fmt.Errorf("dial to `%s` failed: %w", address, err),
+				xerrors.WithName("DialContext"),
+				xerrors.WithBackoff(backoff.TypeSlow),
+				xerrors.WithDeleteSession(),
+			),
+		)
+
+		c.onTransportError(ctx, err)
+
+		return nil, err
 	}
 
 	c.cc = cc
 	c.setState(Online)
 
 	return c.cc, nil
+}
+
+func (c *conn) onTransportError(ctx context.Context, cause error) {
+	for _, onTransportError := range c.onTransportErrors {
+		onTransportError(ctx, c, cause)
+	}
 }
 
 func (c *conn) touchLastUsage() {
@@ -262,8 +289,8 @@ func (c *conn) Close(ctx context.Context) (err error) {
 
 	c.setState(Destroyed)
 
-	for _, f := range c.onClose {
-		f(c)
+	for _, onClose := range c.onClose {
+		onClose(c)
 	}
 
 	return xerrors.WithStackTrace(err)
@@ -301,18 +328,26 @@ func (c *conn) Invoke(
 	c.touchLastUsage()
 	defer c.touchLastUsage()
 
-	err = cc.Invoke(ctx, method, req, res, opts...)
+	ctx, sentMark := markContext(ctx)
 
+	err = cc.Invoke(ctx, method, req, res, opts...)
 	if err != nil {
 		if wrapping {
-			return xerrors.WithStackTrace(
-				xerrors.FromGRPCError(
-					err,
-					xerrors.WithAddress(c.Address()),
-				),
+			err = xerrors.FromGRPCError(err,
+				xerrors.WithAddress(c.Address()),
 			)
+			if sentMark.canRetry() {
+				err = xerrors.Retryable(err,
+					xerrors.WithName("Invoke"),
+					xerrors.WithDeleteSession(),
+				)
+			}
+			err = xerrors.WithStackTrace(err)
 		}
-		return xerrors.WithStackTrace(err)
+
+		c.onTransportError(ctx, err)
+
+		return err
 	}
 
 	if o, ok := res.(response.Response); ok {
@@ -381,24 +416,30 @@ func (c *conn) NewStream(
 	c.touchLastUsage()
 	defer c.touchLastUsage()
 
-	s, err = cc.NewStream(ctx, desc, method, opts...)
+	ctx, sentMark := markContext(ctx)
 
+	s, err = cc.NewStream(ctx, desc, method, opts...)
 	if err != nil {
 		if wrapping {
-			return s, xerrors.WithStackTrace(
-				xerrors.FromGRPCError(
-					err,
+			err = xerrors.Retryable(
+				xerrors.FromGRPCError(err,
 					xerrors.WithAddress(c.Address()),
 				),
+				xerrors.WithName("NewStream"),
+				xerrors.WithDeleteSession(),
 			)
 		}
-		return s, xerrors.WithStackTrace(err)
+
+		c.onTransportError(ctx, err)
+
+		return s, err
 	}
 
 	return &grpcClientStream{
 		ClientStream: s,
 		c:            c,
 		wrapping:     wrapping,
+		sentMark:     sentMark,
 		onDone: func(ctx context.Context) {
 			cancel()
 		},
@@ -416,20 +457,95 @@ func withOnClose(onClose func(*conn)) option {
 	}
 }
 
+func withOnTransportError(onTransportError func(ctx context.Context, cc Conn, cause error)) option {
+	return func(c *conn) {
+		if onTransportError != nil {
+			c.onTransportErrors = append(c.onTransportErrors, onTransportError)
+		}
+	}
+}
+
 func newConn(e endpoint.Endpoint, config Config, opts ...option) *conn {
+	grpcDialOptions := config.GrpcDialOptions()
+	grpcDialOptions = append(
+		append(
+			make([]grpc.DialOption, 0, len(grpcDialOptions)+1),
+			statsHandlerOption,
+		),
+		grpcDialOptions...,
+	)
 	c := &conn{
-		state:    Created,
-		endpoint: e,
-		config:   config,
-		done:     make(chan struct{}),
-		onClose:  make([]func(*conn), 0),
+		grpcDialOptions: grpcDialOptions,
+		state:           Created,
+		endpoint:        e,
+		config:          config,
+		done:            make(chan struct{}),
 	}
 	for _, o := range opts {
 		o(c)
 	}
+
 	return c
 }
 
 func New(e endpoint.Endpoint, config Config, opts ...option) Conn {
 	return newConn(e, config, opts...)
+}
+
+var _ stats.Handler = statsHandler{}
+
+type statsHandler struct{}
+
+var statsHandlerOption = grpc.WithStatsHandler(statsHandler{})
+
+func (statsHandler) TagRPC(ctx context.Context, _ *stats.RPCTagInfo) context.Context {
+	return ctx
+}
+
+func (statsHandler) HandleRPC(ctx context.Context, rpcStats stats.RPCStats) {
+	switch rpcStats.(type) {
+	case *stats.Begin:
+		getContextMark(ctx).markSafeToRetry()
+	default:
+		getContextMark(ctx).markDirty()
+	}
+}
+
+func (statsHandler) TagConn(ctx context.Context, _ *stats.ConnTagInfo) context.Context {
+	return ctx
+}
+
+func (statsHandler) HandleConn(context.Context, stats.ConnStats) {}
+
+type ctxHandleRPCKey struct{}
+
+var rpcKey = ctxHandleRPCKey{}
+
+func markContext(ctx context.Context) (context.Context, *modificationMark) {
+	mark := &modificationMark{}
+	return context.WithValue(ctx, rpcKey, mark), mark
+}
+
+func getContextMark(ctx context.Context) *modificationMark {
+	v := ctx.Value(rpcKey)
+	if v == nil {
+		return &modificationMark{}
+	}
+	return v.(*modificationMark)
+}
+
+type modificationMark struct {
+	safeToRetry uint32
+}
+
+func (m *modificationMark) canRetry() bool {
+	return atomic.LoadUint32(&m.safeToRetry) != 0
+}
+
+func (m *modificationMark) markSafeToRetry() {
+	atomic.StoreUint32(&m.safeToRetry, 1)
+}
+
+func (m *modificationMark) markDirty() {
+	atomic.StoreUint32(&m.safeToRetry, 0)
 }