+++ CHANGELOG.md
@@ -1,6 +1,8 @@
+* Add transport errors checks in retry
+
 ## v3.28.0
 * Added `sugar.GenerateDeclareSection()` helper for make declare section in `YQL`
-* Added check when parameter name not started from `$` and automatically prepends it to name 
+* Added check when parameter name not started from `$` and automatically prepends it to name
 * Refactor connection closing
 
 ## v3.27.0